Correccion de formulario de pago, descuento en productos y filtros en la vista inventario

+++ database/migrations/2021_06_01_082356_create_descuento_table.php
@@ -15,7 +15,7 @@ class CreateDescuentoTable extends Migration
     {
         Schema::create('descuento', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('products_id')->unsigned()->nullable();
+            $table->bigInteger('products_id')->unsigned()->nullable();
             $table->foreign('products_id')->references('id')->on('products');
             $table->date('initial_date')->nullable();
             $table->date('final_date')->nullable();
